Integrate bloc with login screen and add sign out functionality to the main screen view
