+++ frontend/lib/screens/main/bloc/main_screen_event.dart
@@ -3,3 +3,5 @@ abstract class MainScreenEvent {}
 class LoadMainScreen extends MainScreenEvent {}
 
 class GetTasksAndStatistics extends MainScreenEvent {}
+
+class SignOutRequest extends MainScreenEvent {}
